Major updates!!
For this commit I renamed the executable so that it better reflects what it does. Also I updated it so that you can use it for any commit in any project. You can still install it as a hook but that's no longer the only thing it does. The README has updated instructions for how to use it. {BLOG}